@@ -1,6 +1,6 @@
 ﻿namespace ToolBX.Collections.Inventory;
 
-public record StockSearchResult<T> : IReadOnlyList<IndexedEntry<T>>, IEquatable<IEnumerable<IndexedEntry<T>>>
+public sealed record StockSearchResult<T> : IReadOnlyList<IndexedEntry<T>>, IEquatable<IEnumerable<IndexedEntry<T>>>
 {
     private readonly IReadOnlyList<IndexedEntry<T>> _items;
 
@@ -11,39 +11,36 @@ public record StockSearchResult<T> : IReadOnlyList<IndexedEntry<T>>, IEquatable<
     public StockSearchResult(IEnumerable<IndexedEntry<T>> items)
     {
         if (items == null) throw new ArgumentNullException(nameof(items));
-        _items = items as IReadOnlyList<IndexedEntry<T>> ?? items.ToList();
+        _items = items.ToImmutableList();
     }
 
     public IEnumerator<IndexedEntry<T>> GetEnumerator() => _items.GetEnumerator();
 
     IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
-    public virtual bool Equals(StockSearchResult<T>? other) => Equals(other as IEnumerable<IndexedEntry<T>>);
+    public bool Equals(StockSearchResult<T>? other) => Equals(other as IEnumerable<IndexedEntry<T>>);
 
-    public virtual bool Equals(IEnumerable<IndexedEntry<T>>? other)
+    public bool Equals(IEnumerable<IndexedEntry<T>>? other)
     {
         if (other is null) return false;
         if (ReferenceEquals(this, other)) return true;
         return this.SequenceEqual(other);
     }
 
-    public override int GetHashCode()
-    {
-        return _items.GetHashCode();
-    }
+    public override int GetHashCode() => _items.GetValueHashCode();
 
     public IReadOnlyList<GroupedEntry<T>> Group()
     {
         if (Count == 0) return Array.Empty<GroupedEntry<T>>();
 
-        var items = this.DistinctBy(x => x.Item);
+        var distinctEntries = this.DistinctBy(x => x.Item);
         var group = new List<GroupedEntry<T>>();
-        foreach (var item in items)
+        foreach (var entry in distinctEntries)
         {
-            var entries = this.Where(x => Equals(x.Item, item)).ToList();
-            var quantity = entries.Sum(x => x.Quantity);
-            var indexes = entries.Select(x => x.Index);
-            group.Add(new GroupedEntry<T>(item.Item, quantity, indexes));
+            var duplicateEntries = _items.Where(x => Equals(x.Item, entry.Item)).ToList();
+            var quantity = duplicateEntries.Sum(x => x.Quantity);
+            var indexes = duplicateEntries.Select(x => x.Index);
+            group.Add(new GroupedEntry<T>(entry.Item, quantity, indexes));
         }
         return group;
     }
