@@ -2,5 +2,9 @@
 
 public static class ObservableStackExtensions
 {
-    public static ObservableStack<T> ToObservableStack<T>(this IEnumerable<T> source) => new(source);
+    public static ObservableStack<T> ToObservableStack<T>(this IEnumerable<T> source)
+    {
+        if (source is null) throw new ArgumentNullException(nameof(source));
+        return new ObservableStack<T>(source);
+    }
 }
