@@ -2,17 +2,22 @@
 
 public static class ReadOnlyListExtensions
 {
-    public static IReadOnlyList<T> ToReadOnlyList<T>(this IEnumerable<T> source) => new ReadOnlyList<T>(source);
+    public static ReadOnlyList<T> ToReadOnlyList<T>(this IEnumerable<T> source)
+    {
+        if (source is null) throw new ArgumentNullException(nameof(source));
+        if (source is ReadOnlyList<T> readOnlyList) return readOnlyList;
+        return new ReadOnlyList<T>(source);
+    }
 
     /// <summary>
-    /// Returns a new read-only list with the specified elements.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> with the specified elements.
     /// </summary>
-    public static IReadOnlyList<T> With<T>(this IReadOnlyList<T> source, params T[] items) => source.With(items as IEnumerable<T>);
+    public static ReadOnlyList<T> With<T>(this IReadOnlyList<T> source, params T[] items) => source.With(items as IEnumerable<T>);
 
     /// <summary>
-    /// Returns a new read-only list with the specified elements.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> with the specified elements.
     /// </summary>
-    public static IReadOnlyList<T> With<T>(this IReadOnlyList<T> source, IEnumerable<T> items)
+    public static ReadOnlyList<T> With<T>(this IReadOnlyList<T> source, IEnumerable<T> items)
     {
         if (source == null) throw new ArgumentNullException(nameof(source));
         if (items == null) throw new ArgumentNullException(nameof(items));
@@ -20,9 +25,9 @@ public static class ReadOnlyListExtensions
     }
 
     /// <summary>
-    /// Returns a new read-only list without the specified elements.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> without the specified elements.
     /// </summary>
-    public static IReadOnlyList<T> Without<T>(this IReadOnlyList<T> source, Func<T, bool> predicate)
+    public static ReadOnlyList<T> Without<T>(this IReadOnlyList<T> source, Func<T, bool> predicate)
     {
         if (source == null) throw new ArgumentNullException(nameof(source));
         if (predicate == null) throw new ArgumentNullException(nameof(predicate));
@@ -30,29 +35,29 @@ public static class ReadOnlyListExtensions
     }
 
     /// <summary>
-    /// Returns a new read-only list without the specified elements.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> without the specified elements.
     /// </summary>
-    public static IReadOnlyList<T> Without<T>(this IReadOnlyList<T> source, params T[] items) => source.Without(items as IEnumerable<T>);
+    public static ReadOnlyList<T> Without<T>(this IReadOnlyList<T> source, params T[] items) => source.Without(items as IEnumerable<T>);
 
     /// <summary>
-    /// Returns a new read-only list without the specified elements.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> without the specified elements.
     /// </summary>
-    public static IReadOnlyList<T> Without<T>(this IReadOnlyList<T> source, IEnumerable<T> items)
+    public static ReadOnlyList<T> Without<T>(this IReadOnlyList<T> source, IEnumerable<T> items)
     {
         if (source == null) throw new ArgumentNullException(nameof(source));
         if (items == null) throw new ArgumentNullException(nameof(items));
         return new ReadOnlyList<T>(source.Where(x => !items.Contains(x)));
     }
-    
-    /// <summary>
-    /// Returns a new read-only list with the specified elements inserted at index.
-    /// </summary>
-    public static IReadOnlyList<T> WithAt<T>(this IReadOnlyList<T> source, int index, params T[] items) => source.WithAt(index, items as IEnumerable<T>);
 
     /// <summary>
-    /// Returns a new read-only list with the specified elements inserted at index.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> with the specified elements inserted at index.
     /// </summary>
-    public static IReadOnlyList<T> WithAt<T>(this IReadOnlyList<T> source, int index, IEnumerable<T> items)
+    public static ReadOnlyList<T> WithAt<T>(this IReadOnlyList<T> source, int index, params T[] items) => source.WithAt(index, items as IEnumerable<T>);
+
+    /// <summary>
+    /// Returns a new <see cref="ReadOnlyList{T}"/> with the specified elements inserted at index.
+    /// </summary>
+    public static ReadOnlyList<T> WithAt<T>(this IReadOnlyList<T> source, int index, IEnumerable<T> items)
     {
         if (source == null) throw new ArgumentNullException(nameof(source));
         if (items == null) throw new ArgumentNullException(nameof(items));
@@ -62,9 +67,9 @@ public static class ReadOnlyListExtensions
     }
 
     /// <summary>
-    /// Returns a new read-only list without the element at the specified index.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> without the element at the specified index.
     /// </summary>
-    public static IReadOnlyList<T> WithoutAt<T>(this IReadOnlyList<T> source, int index)
+    public static ReadOnlyList<T> WithoutAt<T>(this IReadOnlyList<T> source, int index)
     {
         if (source == null) throw new ArgumentNullException(nameof(source));
         var content = source.ToList();
@@ -73,9 +78,9 @@ public static class ReadOnlyListExtensions
     }
 
     /// <summary>
-    /// Returns a new read-only with the elements at the specified index swapped.
+    /// Returns a new <see cref="ReadOnlyList{T}"/> with the elements at the specified index swapped.
     /// </summary>
-    public static IReadOnlyList<T> WithSwapped<T>(this IReadOnlyList<T> source, int current, int destination)
+    public static ReadOnlyList<T> WithSwapped<T>(this IReadOnlyList<T> source, int current, int destination)
     {
         if (source == null) throw new ArgumentNullException(nameof(source));
         var content = source.ToList();
